Fix kitodo/kitodo-presentation#317 - use better higher resolution
By default, the OpenLayers renderer select the closest resolution if the
current resolution of the view does not exactly match a resolution of the
tile source. This could be a lower resolution leading to a worse image
quality.

To enforce the usage of the better quality / higher resolution tiles, the
`zDirection` option is introduced to the `TileLayer`.